+++ ContentPatcher/Framework/AssetLoader.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.IO;
 using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI;
+using xTile;
+using xTile.Tiles;
 
 namespace ContentPatcher.Framework
 {
@@ -15,6 +17,9 @@ namespace ContentPatcher.Framework
         /// <summary>The textures provided by content packs.</summary>
         /// <remarks>This field is used to circumvent an issue where PNGs can't be loaded while a draw is in progress. Since we can't predict when the game will ask for a texture, we preload them ahead of time.</remarks>
         private readonly IDictionary<string, Texture2D> PngTextureCache = new Dictionary<string, Texture2D>();
+
+        ///// <summary>The maps provided by content packs.</summary>
+        //private readonly IDictionary<string, Map> MapCache = new Dictionary<string, Map>();
 
 
         /*********
@@ -29,16 +34,32 @@ namespace ContentPatcher.Framework
         }
 
         /// <summary>Preload an asset from the content pack if necessary.</summary>
-        /// <typeparam name="T">The asset type.</typeparam>
         /// <param name="contentPack">The content pack.</param>
         /// <param name="key">The asset key.</param>
-        public void PreloadIfNeeded<T>(IContentPack contentPack, string key)
+        public void PreloadIfNeeded(IContentPack contentPack, string key)
         {
-            if (this.IsPng<T>(key))
+            // PNG asset
+            if (this.IsPngPath(key))
             {
                 string actualAssetKey = contentPack.GetActualAssetKey(key);
                 if (!this.PngTextureCache.ContainsKey(actualAssetKey))
                     this.PngTextureCache[actualAssetKey] = contentPack.LoadAsset<Texture2D>(key);
+            }
+
+            // map PNG tilesheets
+            if (this.TryLoadMap(contentPack, key, out Map map))
+            {
+                string relativeRoot = contentPack.GetActualAssetKey(""); // warning: this depends on undocumented SMAPI implementation details
+                foreach (TileSheet tilesheet in map.TileSheets)
+                {
+                    if (!tilesheet.ImageSource.StartsWith(relativeRoot) || Path.GetExtension(tilesheet.ImageSource).Equals(".png", StringComparison.InvariantCultureIgnoreCase))
+                        continue; // not a content pack PNG
+
+                    string relativePath = tilesheet.ImageSource.Substring(relativeRoot.Length + 1);
+                    string actualAssetKey = contentPack.GetActualAssetKey(relativePath);
+                    if (!this.PngTextureCache.ContainsKey(actualAssetKey) && this.AssetExists(contentPack, relativePath))
+                        this.PngTextureCache[actualAssetKey] = contentPack.LoadAsset<Texture2D>(relativePath);
+                }
             }
         }
 
@@ -49,7 +70,7 @@ namespace ContentPatcher.Framework
         public T Load<T>(IContentPack contentPack, string key)
         {
             // load from PNG cache if applicable
-            if (this.IsPng<T>(key))
+            if (typeof(T) == typeof(Texture2D) && this.IsPngPath(key))
             {
                 string actualAssetKey = contentPack.GetActualAssetKey(key);
                 if (this.PngTextureCache.TryGetValue(actualAssetKey, out Texture2D texture))
@@ -65,13 +86,37 @@ namespace ContentPatcher.Framework
         ** Private methods
         *********/
         /// <summary>Get whether an asset is a unpacked PNG file.</summary>
-        /// <typeparam name="T">The asset type.</typeparam>
         /// <param name="key">The asset key in the content pack.</param>
-        private bool IsPng<T>(string key)
+        private bool IsPngPath(string key)
         {
-            return
-                typeof(T) == typeof(Texture2D)
-                && key.EndsWith(".png", StringComparison.InvariantCultureIgnoreCase);
+            return key.EndsWith(".png", StringComparison.InvariantCultureIgnoreCase);
+        }
+
+        /// <summary>Try to load an asset key as a map file.</summary>
+        /// <param name="contentPack">The content pack to load.</param>
+        /// <param name="key">The asset key in the content pack.</param>
+        /// <param name="map">The loaded map.</param>
+        /// <returns>Returns whether the map was successfully loaded.</returns>
+        private bool TryLoadMap(IContentPack contentPack, string key, out Map map)
+        {
+            // ignore if we know it's not a map
+            if (!key.EndsWith(".tbin", StringComparison.InvariantCultureIgnoreCase) && !key.EndsWith(".xnb", StringComparison.InvariantCultureIgnoreCase))
+            {
+                map = null;
+                return false;
+            }
+
+            // try to load map
+            try
+            {
+                map = contentPack.LoadAsset<Map>(key);
+                return true;
+            }
+            catch
+            {
+                map = null;
+                return false;
+            }
         }
     }
 }